moved old load_naics, replaced with new code. added migrations. removed unused sql file
