--- usaspending_api/common/management/sql/naics_add_column.sql
@@ -1,1 +1,0 @@
-ALTER TABLE "public"."naics" ADD COLUMN "year" text;
